Rewriting the CHESS game to RESTful API
Using MongoDB, NodeJS, Express and Mongoose to create a database with the chess board, chess pieces and location coordinates on the border of the board. I've managed to have the information in the chess database, specifically in the chessboard, chess pieces and chessborder collections, displayed as json data in the browser using a local virtual server. I plan on using another collection in the database to log the moves and the turns and all these together to display the current board state using javascript/Angular.

This is my first project, so I may be making mistakes or create incorrect processes or superfluous coding that is not yet very efficient. I use this project as a goal to create a functional website that is interactive, in order for me to learn javascript. I haven't looked at similar web-based chess games applications yet, because I want to learn by trying, not copying. 

The first stage would be to have a chess board presented on the screen (using an HT